[#60] Adding initial implementation of new version
This code includes the visual code for the new version logic. Further
changes are required in the backend to change the "runFilter" call to a
new call that will populate the modal with a list of all the possible
filters and create a new build based on those filters with respect to
each version in it's collection. Please refer to the issue for more
information.